Trait tokio_trace::field::AsField

pub trait AsField: Sealed {
    fn as_field(&self, metadata: &Metadata) -> Option<Field>;
}

Trait implemented to allow a type to be used as a field key.

Note: Although this is implemented for both the [Field] type and any type that can be borrowed as an &str, only Field allows O(1) access. Indexing a field with a string results in an iterative search that performs string comparisons. Thus, if possible, once the key for a field is known, it should be used whenever possible.

Required Methods

fn as_field(&self, metadata: &Metadata) -> Option<Field>

Attempts to convert &self into a Field with the specified metadata.

If metadata defines this field, then the field is returned. Otherwise, this returns None.