Struct tokio_trace::span::Span

pub struct Span<'a> { /* fields omitted */ }

A handle representing a span, with the capability to enter the span if it exists.

If the span was rejected by the current Subscriber's filter, entering the span will silently do nothing. Thus, the handle can be used in the same manner regardless of whether or not the trace is currently being collected.

Methods

impl<'a> Span<'a>

pub fn new(meta: &'a Metadata<'a>, values: &ValueSet) -> Span<'a>

Constructs a new Span with the given metadata and set of field values.

The new span will be constructed by the currently-active Subscriber, with the current span as its parent (if one exists).

After the span is constructed, field values and/or follows_from annotations may be added to it.

pub fn new_disabled() -> Span<'a>

Constructs a new disabled span.

pub fn enter<F: FnOnce() -> T, T>(&mut self, f: F) -> T

Executes the given function in the context of this span.

If this span is enabled, then this function enters the span, invokes and then exits the span. If the span is disabled, f will still be invoked, but in the context of the currently-executing span (if there is one).

Returns the result of evaluating f.

pub fn field<Q>(&self, name: &Q) -> Option<Field> where
Q: Borrow<str>,

Returns a Field for the field with the given name, if one exists,

pub fn has_field<Q: ?Sized>(&self, field: &Q) -> bool where
Q: AsField,

Returns true if this Span has a field for the given Field or field name.

pub fn record<Q: ?Sized, V>(&mut self, field: &Q, value: &V) -> &mut Self where
Q: AsField,
V: Value,

Records that the field described by field has the value value.

pub fn record_all(&mut self, values: &ValueSet) -> &mut Self

Record all the fields in the span

pub fn close(&mut self)

Closes this span handle, dropping its internal state.

Once this function has been called, subsequent calls to enter on this handle will no longer enter the span. If this is the final handle with the potential to enter that span, the subscriber may consider the span to have ended.

pub fn is_closed(&self) -> bool

Returns true if this span is closed.

pub fn is_disabled(&self) -> bool

Returns true if this span was disabled by the subscriber and does not exist.

pub fn follows_from(&self, from: &Id) -> &Self

Indicates that the span with the given ID has an indirect causal relationship with this span.

This relationship differs somewhat from the parent-child relationship: a span may have any number of prior spans, rather than a single one; and spans are not considered to be executing inside of the spans they follow from. This means that a span may close even if subsequent spans that follow from it are still open, and time spent inside of a subsequent span should not be included in the time its precedents were executing. This is used to model causal relationships such as when a single future spawns several related background tasks, et cetera.

If this span is disabled, or the resulting follows-from relationship would be invalid, this function will do nothing.

pub fn id(&self) -> Option<Id>

Returns this span's Id, if it is enabled.

pub fn metadata(&self) -> Option<&'a Metadata<'a>>

Returns this span's Metadata, if it is enabled.